+++ dpkt/__init__.py
@@ -33,6 +33,7 @@ import igmp
 import ip
 import ip6
 import ipx
+import llc
 import loopback
 import mrt
 import netbios
